+++ projects/ruby-3.3.10/ext/openssl/ossl_x509store.c
@@ -934,25 +934,25 @@ Init_ossl_x509store(void)
      *
      * See also the man page X509_STORE_CTX_set_verify_cb(3).
      */
-    rb_attr(cX509Store, rb_intern("verify_callback"), 1, 0, Qfalse);
+    rb_attr(cX509Store, rb_intern("verify_callback"), 1, 0, (int)Qfalse);
     /*
      * The error code set by the last call of #verify.
      *
      * See also StoreContext#error.
      */
-    rb_attr(cX509Store, rb_intern("error"), 1, 0, Qfalse);
+    rb_attr(cX509Store, rb_intern("error"), 1, 0, (int)Qfalse);
     /*
      * The description for the error code set by the last call of #verify.
      *
      * See also StoreContext#error_string.
      */
-    rb_attr(cX509Store, rb_intern("error_string"), 1, 0, Qfalse);
+    rb_attr(cX509Store, rb_intern("error_string"), 1, 0, (int)Qfalse);
     /*
      * The certificate chain constructed by the last call of #verify.
      *
      * See also StoreContext#chain.
      */
-    rb_attr(cX509Store, rb_intern("chain"), 1, 0, Qfalse);
+    rb_attr(cX509Store, rb_intern("chain"), 1, 0, (int)Qfalse);
     rb_define_alloc_func(cX509Store, ossl_x509store_alloc);
     rb_define_method(cX509Store, "initialize",   ossl_x509store_initialize, -1);
     rb_undef_method(cX509Store, "initialize_copy");
